requirements.analyzer.txt
streamlit>=1.24.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=0.8.11 